Change the mallet crf tagger to crfsuite crf tagger

+++ nltk/tag/crfsuite.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-# Natural Language Toolkit: Interface to the Stanford NER-tagger
+# Natural Language Toolkit: Interface to the CRFSuite Tagger
 #
 # Copyright (C) 2001-2015 NLTK Project
 # Author: Long Duong <longdt219@gmail.com>
@@ -7,63 +7,47 @@
 # For license information, see LICENSE.TXT
 
 """
-A module for interfacing with the Mallet CRF taggers.
+A module for interfacing with the CRFSuite taggers.
 """
-
+from __future__ import absolute_import
+from __future__ import unicode_literals
 import os
 import tempfile
 from subprocess import PIPE
 import unicodedata
 import re 
-
-from nltk.internals import find_jar, config_java, java, _java_options
+from subprocess import Popen
 from nltk.tag.api import TaggerI
 
-_mallet_url = 'http://mallet.cs.umass.edu/grmm/'
+_crfsuite_url = 'http://www.chokkan.org/software/crfsuite'
 
-class MalletTagger(TaggerI):
+class CRFTagger(TaggerI):
     """
-    An interface to Mallet CRF taggers.
+    An interface to CRFSuite taggers. http://www.chokkan.org/software/crfsuite/tutorial.html
     
-    >>> from nltk.tag.mallet import MalletTagger
-    >>> mt = MalletTagger('/Users/HeroAthen/Dropbox/Work/nltk/temp/mallet-2.0.7/dist/mallet.jar', 
-    ... '/Users/HeroAthen/Dropbox/Work/nltk/temp/mallet-2.0.7/dist/mallet-deps.jar')
+    >>> from nltk.tag.crfsuite import CRFTagger
+    >>> ct = CRFTagger()
  
     >>> train_data = [[('University','Noun'), ('is','Verb'), ('a','Det'), ('good','Adj'), ('place','Noun')],
     ... [('dog','Noun'),('eat','Verb'),('meat','Noun')]]
     
-    >>> mt.train(train_data,'model.crf.tagger')
-    >>> mt.tag_sents([['dog','is','good'], ['Cat','eat','meat']])
-    [[('dog', 'Noun'), ('is', 'Verb'), ('good', 'Noun')], [('Cat', 'Noun'), ('eat', 'Verb'), ('meat', 'Noun')]]
+    >>> ct.train(train_data,'model.crf.tagger')
+    >>> ct.tag_sents([['dog','is','good'], ['Cat','eat','meat']])
+    [[('dog', 'Noun'), ('is', 'Verb'), ('good', 'Adj')], [('Cat', 'Noun'), ('eat', 'Verb'), ('meat', 'Noun')]]
     
     >>> gold_sentences = [[('dog','Noun'),('is','Verb'),('good','Adj')] , [('Cat','Noun'),('eat','Verb'), ('meat','Noun')]] 
-    >>> mt.evaluate(gold_sentences) - 0.83333 < 0.0001
-    True
+    >>> ct.evaluate(gold_sentences) 
+    1.0
     
-    Test the function of finding jar file in the environment variable and setting learned model file  
-    >>> mt = MalletTagger() 
-    >>> mt.set_model_file('model.crf.tagger')
-    >>> mt.evaluate(gold_sentences) - 0.83333 < 0.0001
-    True 
+    Setting learned model file  
+    >>> ct = CRFTagger() 
+    >>> ct.set_model_file('model.crf.tagger')
+    >>> ct.evaluate(gold_sentences)
+    1.0
     
     """
     
-    _JAR = 'mallet.jar'
-    _JAR_DEPS = 'mallet-deps.jar'
-
-    def __init__(self, path_to_mallet_jar=None, path_to_mallet_deps_jar = None, verbose=False, java_options='-mx1000m'):
-
-        self._mallet_jar = find_jar(
-                self._JAR, path_to_mallet_jar,
-                env_vars=('MALLET',), searchpath=(), url=_mallet_url,
-                verbose=verbose)
-        
-        self._mallet_dep_jar = find_jar(
-                self._JAR_DEPS, path_to_mallet_deps_jar,
-                env_vars=('MALLET_DEPS',), searchpath=(), url=_mallet_url,
-                verbose=verbose) 
-        
-        self.java_options = java_options
+    def __init__(self, cmd_options=''):
         self._model_file = ''
     
     def set_model_file(self, model_file):
@@ -79,36 +63,30 @@ class MalletTagger(TaggerI):
              - Suffixes up to length 3 
         :return : a string which contains the features   
         
-        >>> from nltk.tag.mallet import MalletTagger
-        >>> mt = MalletTagger()
-        >>> mt._get_features(u'University1')
-        u'WORD_University1 CAPITALIZATION HAS_NUM SUF_1 SUF_y1 SUF_ty1'
-        >>> mt._get_features(u'.?')
-        u'WORD_.? PUNCTUATION SUF_?'
         """ 
-        feature_list = 'WORD_' + data + ' ' 
+        feature_list = ''  
         # Capitalization 
         if data[0].isupper():
-            feature_list += 'CAPITALIZATION '
+            feature_list += 'CAPITALIZATION\t'
         
         # Number 
         pattern = re.compile('\\d')
         if re.search(pattern, data) is not None:
-            feature_list += 'HAS_NUM ' 
+            feature_list += 'HAS_NUM\t' 
         
         # Punctuation
         punc_cat = set(["Pc", "Pd", "Ps", "Pe", "Pi", "Pf", "Po"])
         if all (unicodedata.category(x) in punc_cat for x in data):
-            feature_list += 'PUNCTUATION '
+            feature_list += 'PUNCTUATION\t'
         
         # Suffix up to length 3
         if len(data) > 1:
-            feature_list += ('SUF_' + data[-1:] + ' ') 
+            feature_list += ('SUF_' + data[-1:] + '\t') 
         if len(data) > 2: 
-            feature_list += ('SUF_' + data[-2:] + ' ')    
+            feature_list += ('SUF_' + data[-2:] + '\t')    
         if len(data) > 3: 
-            feature_list += ('SUF_' + data[-3:] + ' ')
-        
+            feature_list += ('SUF_' + data[-3:] + '\t')
+        feature_list +=  'WORD_' + data + '\t'
         return feature_list.strip()
         
     def tag_sents(self, sents):
@@ -127,40 +105,41 @@ class MalletTagger(TaggerI):
         
         # First, build the test file 
         input_file = tempfile.NamedTemporaryFile(
-                prefix='mallet_tagger.test',
+                prefix='crf_tagger.test',
                 dir=tempfile.gettempdir(),
                 delete=False)
 
         for sent in sentences: 
             for token in sent:
-                data = unicode(token)
-                input_file.write(self._get_features(data) + '\n')
-            input_file.write('\n')
+                #data = unicode(token)
+                data = token
+                input_file.write(('DUMMY_LABEL\t' + self._get_features(data) + '\n').encode('utf-8'))
+            input_file.write('\n'.encode('utf-8'))
         
         input_file.close()
         
         # Now use the model to tag
+        _crf_cmd = ['crfsuite', 'tag', '-m', self._model_file, input_file.name]
+            
+            
+        # Run the tagger and get the output
+        p = Popen(_crf_cmd, stdin=PIPE, stdout=PIPE, stderr=PIPE)
+        (stdout, stderr) = p.communicate()
+
+        # Check the return code.
+        if p.returncode != 0:
+            raise Exception('CRFSuite command failed! Details: %s' % stderr)
         
-        default_options = ' '.join(_java_options)
-        config_java(options=self.java_options, verbose=False)
-        
-        self._cmd = ['cc.mallet.fst.SimpleTagger', '--model-file', self._model_file, input_file.name ]
-        
-        class_path = self._mallet_dep_jar + ':' + self._mallet_jar 
-        stand_output, _stderr = java(self._cmd,classpath=class_path, stdout=PIPE, stderr=PIPE)
-        
-        # Return java configurations to their default values
-        config_java(options=default_options, verbose=False)
-        
-        # Remove the temp file 
+        # Remove the temp file
+        #print ('Test file : ' + input_file.name) 
         os.remove(input_file.name)
         
-        return self.parse_output(stand_output, sentences)
+        return self.parse_output(stdout, sentences)
         
     
     def train(self, train_data, model_file):
         '''
-        Train the CRF tagger using Mallet  
+        Train the CRF tagger using CRFSuite  
         :params train : is the list of annotated sentences.        
         :type train : list (list(tuple(str,str)))
         :params model_file : the model will be saved to this file.     
@@ -169,34 +148,34 @@ class MalletTagger(TaggerI):
         
         try:
             input_file = tempfile.NamedTemporaryFile(
-                prefix='mallet_tagger.train',
+                prefix='crf_tagger.train',
                 dir=tempfile.gettempdir(),
                 delete=False)
 
             for sent in train_data:
                 for data,label in sent:
-                    data = unicode(data)
-                    input_file.write(self._get_features(data) + ' ' + label + '\n')
-                input_file.write('\n')            
+                    #data = unicode(data)
+                    input_file.write((label + '\t' + self._get_features(data) + '\n').encode('utf-8'))
+                input_file.write('\n'.encode('utf-8'))            
             input_file.close()
             
-            # Now train the model
-            default_options = ' '.join(_java_options)
-            config_java(options=self.java_options, verbose=False)
+            # Now train the model, the output should be model_file
+            _crf_cmd = ['crfsuite', 'learn', '-m', model_file, input_file.name]
             
-            self._cmd = ['cc.mallet.fst.SimpleTagger',
-                '--train', 'true', '--model-file', model_file, input_file.name ]
-            
-            # Run the java command
-            class_path = self._mallet_dep_jar + ':' + self._mallet_jar 
-            stand_output, _stderr = java(self._cmd,classpath=class_path, stdout=PIPE, stderr=PIPE)
-            # Return java configurations to their default values
-            config_java(options=default_options, verbose=False)
+            # Serialize the actual sentences to a temporary string
+            # Run the tagger and get the output
+            p = Popen(_crf_cmd, stdin=PIPE, stdout=PIPE, stderr=PIPE)
+            (stdout, stderr) = p.communicate()
+
+            # Check the return code.
+            if p.returncode != 0:
+                raise Exception('CRFSuite command failed! Details: %s' % stderr)
             
             # Save the model file 
             self._model_file = model_file            
         finally:
             os.remove(input_file.name)
+            #print ('Training data : ' + input_file.name)
 
     def tag(self, tokens):
         '''
@@ -215,6 +194,7 @@ class MalletTagger(TaggerI):
         
         labels= []
         sent_labels = []
+        text = text.decode('utf-8')
         for label in text.strip().split("\n"):
             label = label.strip()
             if label == '':
@@ -228,7 +208,7 @@ class MalletTagger(TaggerI):
             
         # Match labels with word 
         if len(sentences) != len(sent_labels):
-            raise ValueError(' Expecting error, number of sentence is not matched')
+            raise ValueError(' Expecting error, number of sentence is not matched' + str(len(sentences)) + ' vs ' + str(len(sent_labels)))
         tagged_sentences = []
          
         for i  in range(len(sentences)):
